made a new package to organize server side stuff
