+++ src/main/java/org/project/Menu.java
@@ -1,5 +1,9 @@
 package org.project;
 
+import org.project.ServerSide.Client;
+import org.project.ServerSide.SaveState;
+import org.project.ServerSide.SaveStateHandler;
+
 import javax.swing.*;
 import java.awt.*;
 import java.awt.event.ActionListener;
