+++ src/main/java/org/project/ServerSide/PostRequestHandler.java
@@ -1,7 +1,9 @@
-package org.project;
+package org.project.ServerSide;
 
 import org.bson.Document;
 import org.json.simple.*;
+import org.project.ServerSide.DatabaseHandler;
+
 import java.io.*;
 import java.net.Socket;
