+++ src/main/java/org/project/ServerSide/SaveState.java
@@ -1,7 +1,8 @@
-package org.project;
+package org.project.ServerSide;
 
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
+import org.project.*;
 import org.project.Entities.Entity;
 import org.project.Entities.Monster;
 import org.project.Entities.Player;
