+++ src/test/java/org/project/SaveStateHandlerTest.java
@@ -2,6 +2,7 @@ package org.project;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.project.ServerSide.SaveState;
 
 public class SaveStateHandlerTest {
   private GamePanel gamePanel;
